ADD Refactor to change the new name and prepare publishing
