+++ config.js
@@ -5,7 +5,7 @@ var config = {};
 config.server = {
     port: 60001,                         // Port where the server will be listening
     defaultType: 'Device',
-    logLevel: 'DEBUG'
+    logLevel: 'FATAL'
 };
 
 // Configuration of the LWTM2M Client
