+++ lib/lwm2m-client.js
