+++ lib/lwm2m-node-lib.js
@@ -23,5 +23,5 @@
 
 'use strict';
 
-exports.server = require('./iotagent-lwm2m-server');
-exports.client = require('./iotagent-lwm2m-client');
+exports.server = require('./lwm2m-server');
+exports.client = require('./lwm2m-client');
