+++ lib/lwm2m-server.js
