+++ package.json
@@ -1,20 +1,20 @@
 {
-  "name": "iotagent-lwm2m-lib",
-  "description": "Library for Lightweight M2M IoT Agents.",
+  "name": "lwm2m-node-lib",
+  "description": "Library for developing servers and client of OMA Lightweight M2M",
   "version": "0.2.0",
-  "homepage": "https://github.com/telefonicaid/iotagent-lwm2m-lib",
+  "homepage": "https://github.com/telefonicaid/lwm2m-node-lib",
   "author": {
     "name": "Daniel Moran",
     "email": "daniel.moranjimenez@telefonica.com"
   },
   "repository": {
     "type": "git",
-    "url": "git://github.com/telefonicaid/iotagent-lwm2m-lib.git"
+    "url": "git://github.com/telefonicaid/lwm2m-node-lib"
   },
   "bugs": {
-    "url": "https://github.com/telefonicaid/iotagent-lwm2m-lib/issues"
+    "url": "https://github.com/telefonicaid/lwm2m-node-lib/issues"
   },
-  "main": "lib/iotagent-lwm2m-lib",
+  "main": "lib/lwm2m-node-lib",
   "engines": {
     "node": ">=0.10.8"
   },
